# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API pour OC elites du congo",
        "description": "Oc elites du congo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://elite.com/api"
        }
    ],
    "paths": {
...
        "/admin/galeries/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__GALERIE"
                ],
                "summary": "Ajouter des images à la galérie",
                "operationId": "4a40f7a9a7e03da124da037e1413c921",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/galeries/update/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__GALERIE"
                ],
                "summary": "Moofier une image dans la galérie",
                "operationId": "236117eb29df1827df7e0829d0d6e6bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/galeries/delete/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__GALERIE"
                ],
                "summary": "Supprimer une photo liée à un évènement",
                "operationId": "18d0a8c3ae32577ddf5a44cb17a8f4a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/jerseys/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__MAILLOTS"
                ],
                "summary": "Ajouter un maillot",
                "operationId": "078e4f5fdabbc2c376e9f82e31a04e62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Maillot domicile"
                                    },
                                    "chemin": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/jerseys/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__MAILLOTS"
                ],
                "summary": "Renvoie les informations d'une jersey par son ID",
                "operationId": "8b707598825dc3ca02c8555cd6307fa0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/jerseys/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__MAILLOTS"
                ],
                "summary": "Edite les informations d'une jersey",
                "operationId": "9ddd3034c3a0ed5c64fbc8faf1231bbc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Nike"
                                    },
                                    "chemin": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/jerseys/delete/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__MAILLOTS"
                ],
                "summary": "Supprime une jersey par son ID",
                "operationId": "b64d5aeaddb0d46f8568c528f88812d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": [
            []
        ]
    }
}